+++ packages/core/src/index.ts
@@ -2,3 +2,5 @@ export * from "./Graph";
 export * from "./types";
 export * from "./GraphAnalytics";
 export * from "./CommunityDetection";
+export * from "./GraphImporters";
+export * from "./GraphCollaboration";